suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@35475 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/socket/socket.c
@@ -1426,7 +1426,7 @@ sock_s_pack_sockaddr_un(VALUE self, VALUE path)
     StringValue(path);
     MEMZERO(&sockaddr, struct sockaddr_un, 1);
     sockaddr.sun_family = AF_UNIX;
-    if (sizeof(sockaddr.sun_path) <= RSTRING_LEN(path)) {
+    if (sizeof(sockaddr.sun_path) <= (size_t)RSTRING_LEN(path)) {
         rb_raise(rb_eArgError, "too long unix socket path (%"PRIuSIZE" bytes given but %"PRIuSIZE" bytes max)",
             (size_t)RSTRING_LEN(path), sizeof(sockaddr.sun_path));
     }
